+++ guy.py
@@ -22,15 +22,19 @@
 # logger for each part
 # cookiejar
 
-__version__="0.5.7" #autoreload's version !
+__version__="0.6.0" #the one version
 
-import os,sys,re,traceback,copy,types
+import os,sys,re,traceback,copy,types,shutil
 from urllib.parse import urlparse
 from threading import Thread
 import tornado.web
 import tornado.websocket
 import tornado.platform.asyncio
 import tornado.autoreload
+import tornado.httpclient
+from tornado.websocket import websocket_connect
+from tornado.ioloop import IOLoop
+from tornado import gen
 import platform
 import json
 import asyncio
@@ -49,15 +53,16 @@ import io
 class FULLSCREEN: pass
 ISANDROID = "android" in sys.executable
 FOLDERSTATIC="static"
+CHROMECACHE=".cache"
 class JSException(Exception): pass
 
 handler = logging.StreamHandler()
 handler.setFormatter( logging.Formatter('-%(asctime)s %(name)s [%(levelname)s]: %(message)s') )
+handler.setLevel(logging.ERROR)
 logger = logging.getLogger("guy")
 logger.addHandler(handler)
-logger.setLevel(logging.DEBUG)
+logger.setLevel(logging.ERROR)
 
-handler.setLevel(logging.ERROR)
 
 def isFree(ip, port):
     s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -81,11 +86,33 @@ async def callhttp(web,path): # web: RequestHandler
             else:
                 ret=method(web,*g.groups())
             if isinstance(ret,Guy):
+                print("*** DEPRECATED ***, will be removed in near future")
                 ret.parent = web.instance
                 web.write( ret._renderHtml() )
             return True
 #=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#=#
 
+def wsquery(wsurl,msg): # Synchrone call, with tornado
+    """ In a simple world, could be (with websocket_client pypi):
+            ws = websocket.create_connection(wsurl)
+            ws.send(msg)
+            resp=ws.recv()
+            ws.close() 
+            return resp            
+    """
+    @gen.coroutine
+    def fct(ioloop,u,content):
+        cnx=yield websocket_connect(u)
+        cnx.write_message(msg)
+        resp=yield cnx.read_message()
+        cnx.close()
+        ioloop.stop()
+        ioloop.response=resp
+    
+    ioloop = IOLoop.instance()
+    fct(ioloop,wsurl,msg)
+    ioloop.start()
+    return ioloop.response
 
 class readTextFile(str):
     filename=None
@@ -178,6 +205,12 @@ class GuyJSHandler(tornado.web.RequestHandler):
             raise tornado.web.HTTPError(status_code=404)
 
 INST={}
+
+class FavIconHandler(tornado.web.RequestHandler):
+    def initialize(self, instance):
+        self.instance=instance
+    async def get(self):
+        self.write(b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00@\x00\x00\x00@\x08\x06\x00\x00\x00\xaaiq\xde\x00\x00\x00\x06bKGD\x00\xd4\x00{\x00\xff\xf0\x90\n\xda\x00\x00\x00\tpHYs\x00\x00\x0b\x13\x00\x00\x0b\x13\x01\x00\x9a\x9c\x18\x00\x00\x00\x07tIME\x07\xe4\x04\x0e\x0f)\x02\xf5J\x9b=\x00\x00\x00\x1diTXtComment\x00\x00\x00\x00\x00Created with GIMPd.e\x07\x00\x00\x06RIDATx\xda\xdd\x9b_\x88Te\x14\xc0\x7f\xe7\x9bu\xcd\x08D\x82\xfe=\xf4\x92\x88\xceV\xfe\x99\x9d\x82z(0{\x8a\x08\xa2\x07\xa9\xb0Q\x14\xb7\xb2\xcdXVLb\xd42)\x82\xb0\x82\xc5\xd5\x1a,\xd9\xe7\x02{)\x08|*[w\xdd\x9d\xd04\t\x0c\xa2\x1e\x82\xd0(1\xd7\x9d\xd3\xc3\xfc\xd9;3\xf7\xde\xef\xbbw\xef\xfc\xd9.\x0c\xbb;\xf7\xfbs\xbe\xdfw\xce\xf9\xce9\xf7\xae\xd0\x82\xeb\xccKz\'p\x0fp7p\x17p\xbb\x96\xb8\x15X\n\xdc\x02\xdc\x0c\xf4\x02=\x80\x00\x8arCK\\\x07\xae\x02\x7f\x03\x7f\x01\x7f\x02\x7f\x00\xbf\x03\xbf\x02\x972\xa3\xf2\xb3w\xae\xf1MJ\xf6\x98\xc4\x965V\xcf\xd3[\x95\xfe#\xe5\xaegvh\x1f%\xb2\xc0\xa3\xc0\xc3\xc0r\xbf>\xaa\x80Z\x06V\xd0\x92\x83\x00%\xae\x00\xdf\x02\'\x81S\xc0\xf7\x99\xa3\xf2O[\x00L\rj/\xb0^K\xe4\x80g<BY/um\xa3N\x10\x1a\x01\x9fG9\x0c|\xd9\xff\x89\\\x048\x9dS\xfa\x0b\x92\x0c\x80\xa9A]\x05\x0c\x03/\x04\xeeh\xe7!T\xfb\xfd\x06\xbc\xcc\x0c_\xf4\x7f&\xa5\xf1\x9c\x92\r\x00!!\x0bf\xcd!ajPW\x02c\xc0Z\'\xb5\xee4\x84f(\xcfg\x0br<\xb2\x06L\xbf\xaa\x8bU9\n<\x17*\xe8\xc2\x800\x0b\xdc\x9b-\xc8\xf9Fm0u\x8b\xde\xa9\xd5\x9f+\x10\xae\x89\t_<\x80\x88\x0fF\xe3\xe0|\\\xdbHtO\xd68\xb6\x08)\xe0\xc7\xf1\x9cn\xcd\x16\x84\xf1\x9c6\x8bZ\x1cRV\xbf/L\xef\xd4\r\xc0\x85H\x82\xfa\t\x99$\x04\x17\x00v\x08\x00\xa3\xe39=\xe0\x85P\xd7m\xfa5\xdd\x00|\xe5\xa7v\x1aW\xad\x932\x87Y\x07\x10>\xe6\xe8\x1d[\xe7\xee\xbd\x98-\xc8H\x1d\x80\xe2\x90.\x07.\xd6:\xfc\xff!\xac\xcc\x16\xe4\x82)\x0ekUeO\xd5\xa9\x8e$\xa8\xb2\x92\x909\xa4\x9a\xbe\x9a\x01\x0e\x00kQ\x1e\x00>\xb4\x99\x83\xc7\\?\xae\x89V\x1c\xd6-\xc0Q\xefn-\x00M\xc8\xd3\xcb\xfe\xccG\xf5\xab\x9d\xd8\xa6\x068\x82\xb2\xd9A\x13VT\xd9|\xd0\xb8[-\xd1\x04\x93\x98&\xe43\xa3\xb2\x9f\x99\xe6{\x99Q)eFe\x0b\xc2X\xd8\xd8" \xc2\x93\xa68\xac\x0fU\x92\x93&\x8f\xda\xa5\x10~Z7"\xfb\'\x07\x94\xcca\x7f\xdb\x9a\x1cP2\xa3\xf2,\x86k\x96\xb1\xef3\xc0\xe3a\xc7J\x17B(\x00\xac\x1b\tv,\x9e{\x9f\xfb\xcd\xe5\x19\xfb&\x03\xac\xb6\x9d\xad]\x06\xe1d\x84\xdcm,h\xae\xca\xd87\x0cp\x9bK\x80\x11\x06!\xb6\x97\x8f\xd7o&\x02\x80\x7f\xc3\x80\x8b)\x7f=\xe3\x1ae\xd5\x8e\x10q\x08\x87]\x16\xe3\xd8\xaf\xa1\xcd\x83\x93\x03\xf6\x04ab\x9b\x02<e\xd1:5\x92r *s\x9fN@h\x18;\x17f\xff\x9e\xd3\x00`\x93\xc5\xf4\xd4\x04\x04\x18\xdd\x0c!39\xa0\x1b-;\xcf\xc46\xddSw\xba\x05@0!QV(\x04:\x0balr@\x9f\x06\x98\xd8\xaeM;?9\xa0;\x80\xb7B\x07+\xcf\xaf\xf2\xc3n\xfd\x1ax,R\xbc]\xf2\xa9\x05h\xf4\x1a\xa0\xfaE\x87.\xfd\xe6\xda\x9c\x04F\x80\xe9J\x81\xb5\x1f\xd8\x01\xacs\xac/\x1e\xeb\xf1\x8b\xb7\xad\x10LYp\x91\x8a\x1cJ\xb5\xb6[\xb7[\xaav\x85R\xd3\x00A\xec\x00<c?R\xf9\x04kk\xf8Xj\x1c\x93\x8e@;\x920s0nV\xd5\xc9\x82\x8a\x99W!")\x08\x9d\xab*\xa9\t\xdb\x9a\xc8\x10Xx\x10\x8cM?#A0\x0b\x0e\x82\x1a\'#\x956Ch_}Q]\xf6\xd7\xbd:k\xec\xc2tA-\xa1\x0e@\x8fkV!\xc6\xb1N_="\xab\xed}\x8e\xa2\xda\xbd\xf2u9t>"\x1e\x91\xc6~\xfe{\x8e\xfa\xab\xce\x00Z\x04av\xcd!YF\x07/\x13\xb5Cls\x08\xa8%L\r*\x0b\n@\xd2\x10:}\x191\x01\x9e7\x01\x87\xe3]\xb4+\x84\xe2P{5\xa2\xa7v\xdec\x8f\xdd}\x9d\x89\xcd\'\xc8\\\x92\xe3\xe7\x13\x1aA\xde\xff\x9eP\x1c\xd2\xab\x18\xae;\xc9\xe0\xf5/\x12\x96u\xf9:GO2\x14\x17\x82\x99?\x04\x1f\xd3Y\x02,q\x99_\x03\xb2\xd1\xc0\x0c\xb6\xfey\xc7\x12\xd3\x14\xf9\xb5\xca\x1c\x82\xea\x8bI\xf9\x05\x89\xe1\x8f\x04LS\xc7\xb8\x10R\x11!t\xc2!\xfa\x04j\xc6w\x17b\n\x17\xbb\xb4\x96\x90\xa6\xc5\xc9[L\x90*\x8ai\x0b\x04\x89t\x9a\x88e\xfc\x18\xc9\x9b\tKz\xda\x00a&\xce\xee[!D\x88Q\x8c-\xf3k1\x04\xb5\x86b\x01\xbb\x1e*\x97q\x87`\\\xaa\xb3-\x85\x10\xd7\xee\xc5\x0e\xc1\xc5\x97\x19\xdfI\xba\x04\x82\xb5\xaf\x84\x07\xf3\x91j\x82\xdd\x06\xc1\xb5\x8f\xcc\x13\x82\t]`\x92\x10L\xeb\x80\xcd\x07\x82\xb1\n*\xf3[L\xe4#*\xae\xb9\x88\xc3\x06\x88?\x80\x1b\xd6\xb0\xd2D\x9f\xb0\x13\x10\xe2\xa4\xf1=\x94\xdf\xcboZ\x9c6&\x18~Op\\\x92\x90\x00\x08\x95\xa4D\x80\xa5\xc5a\xf5\xaa\xfd\x95VC\xf0>\xf9\xea\xa1\xfcf5q \xf8\xb6\x89\x10\xd5\xa1,\x06.\'q\\FN\xe3g\xe7\x96t!\xd4\xae,\xe6 13\xba\xaa:\xb6{\xf1\x8d\x8e\xd6\x00\xe3V\xe7\xd2J\x08\x1d\xbc$\x05\xa6o\x9fL8y\xd8\x16AH4\xe7\x8f\x9b!K\x8a\xe3\xb1\x8e\x996C\x10K\xa2\x13G\xa3*\xefIp(\xa9\'6-\x83\xe0XE\x8a\n\xc1\x00\xa4\xf3r\x1a\xf8.I\x08IV|\xc4\xd4\xafY\x12\x84\xe0m\xba\xd1\xb5\xb3\xd3\xb3\xbb\x84 \x04\xc9\x93\x14\x84Z\xb3t^.\x01\x9b\xbb\t\x82-,o\x82\x10#o\xa959\xf7\xa6\x92\xceK\x01x\xb7\x1b \x04\xbeY\x1a\x06!F\xf26\xa7\x01o\x08\xe7\xf6)\xe9\xbc\xec\x02\xf6\xcc\x0b\x82_\xa5y\xbe\x89\x8d\xb1@\x88\x99\xbc\xd5\xddJ\xe7\x85\xb3{\x95t^\xde\x06\xd6\xc7\x86\x10\xf3M\x0f\xd7\x98\xc3\xba\xd3\xc6\x1dB\xd3\xd7}{\x85\xf3\xef\x94H\xe7\xe5\x1bJ\xf4\x02\x07\x93\x82`-kG\x01g;n\x1d!\x98\xdd\x9ch\xfar\xe5\xaeZ\xc9t&\x9d\x97\xd7)\xbfr\xfa\n\xf0K\xab 8G\x9b!\xe3\x8bD\x87 \x00\xbb9!\x07y"0\xa7;\x9bW\xfa\xf6I\xf5\xf7;*\xe6\x91\x05VQ\xfe\x17\xf9e\xc0\xa2j{ux\xfb\xb3\xb1M\x9d`\n\xa4\xea\xfe\x9a\x05\xae\x89!\x05\xa4\x9a\xc6\xaf\xfc-=,\xd2RSnj\xfc\xdeF\xad<S\xfc\xf4?\xd9\xf1\xf7\xdeE\\\xb8\xa0\x00\x00\x00\x00IEND\xaeB`\x82')
 
 class MainHandler(tornado.web.RequestHandler):
 
@@ -302,10 +335,11 @@ class WebSocketHandler(tornado.websocket.WebSocketHandler):
         del WebSocketHandler.clients[self]
 
     async def on_message(self, message):
+      
         instance = WebSocketHandler.clients.get(self,None)
         if instance is None:
             return
-
+          
         o = jLoads(message)
         logger.debug("WS RECEPT: %s",o)
         method,args,uuid = o["command"],o.get("args"),o["uuid"]
@@ -400,6 +434,7 @@ class WebServer(Thread): # the webserver is ran on a separated thread
             (r'/(?P<id>[^/]+)-ws',          WebSocketHandler,dict(instance=self.instance)),
             (r'/(?P<id>[^/]+)-js',          GuyJSHandler,dict(instance=self.instance)),
             (r'/(?P<page>[^\.]*)',          MainHandler,dict(instance=self.instance)),
+            (r'/favicon.ico',               FavIconHandler,dict(instance=self.instance)),
             (r'/(.*)',                      tornado.web.StaticFileHandler, dict(path=statics ))
         ])
         self.app.listen(self.port,address=self.host)
@@ -432,9 +467,9 @@ class WebServer(Thread): # the webserver is ran on a separated thread
         return "http://localhost:%s/#%s" % (self.port,self.instance._name) #anchor is important ! (to uniqify ressource in webbrowser)
 
 
+
 class ChromeApp:
-    def __init__(self, url, size=None, chromeArgs=[]):
-        self.__instance = None
+    def __init__(self, url, appname="driver",size=None,lockPort=None,chromeargs=[]):
 
         def find_chrome_win():
             import winreg  # TODO: pip3 install winreg
@@ -465,30 +500,81 @@ class ChromeApp:
                 except webbrowser.Error:
                     exe = None
 
-        if exe:
-            args = [exe, "--app=" + url] + chromeArgs
-            # args.append("--aggressive-cache-discard")
-            if size == FULLSCREEN:
-                args.append("--start-fullscreen")
-            if tempfile.gettempdir():
-                args.append(
-                    "--user-data-dir=%s"
-                    % os.path.join(tempfile.gettempdir(), ".guyapp_"+re.sub(r"[^a-zA-Z]","_",url))
-                )
-            logger.debug("CHROME APP-MODE: %s",args)
-            self.__instance = subprocess.Popen(args)
-        else:
+        if not exe:
             raise Exception("no chrome browser, no app-mode !")
+        else:
+            args = [ #https://peter.sh/experiments/chromium-command-line-switches/
+                exe,
+                "--app=" + url, # need to be a real http page !
+                "--app-id=%s" % (appname),
+                "--app-auto-launched",
+                "--no-first-run",
+                "--no-default-browser-check",
+                "--disable-notifications",
+                "--disable-features=TranslateUI",
+                #~ "--no-proxy-server",
+            ] + chromeargs
+            if size:
+                if size == FULLSCREEN:
+                    args.append("--start-fullscreen")
+                else:
+                    args.append( "--window-size=%s,%s" % (size[0],size[1]) )
+
+            if lockPort: #enable reusable cache folder (coz only one instance can be runned)
+                self.cacheFolderToRemove=None
+                args.append("--remote-debugging-port=%s" % lockPort)
+                args.append("--disk-cache-dir=%s" % CHROMECACHE)
+                args.append("--user-data-dir=%s/%s" % (CHROMECACHE,appname))
+            else:
+                self.cacheFolderToRemove=os.path.join(tempfile.gettempdir(),appname+"_"+str(os.getpid()))
+                args.append("--user-data-dir=" + self.cacheFolderToRemove) 
+                args.append("--aggressive-cache-discard")
+                args.append("--disable-cache")
+                args.append("--disable-application-cache")
+                args.append("--disable-offline-load-stale-cache")
+                args.append("--disk-cache-size=0")
+
+            logger.debug("CHROME APP-MODE: %s"," ".join(args))
+            # self._p = subprocess.Popen(args)
+            self._p = subprocess.Popen(args,stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+
+            #~ if lockPort:
+                #~ http_client = tornado.httpclient.HTTPClient()
+                #~ self._ws = None
+                #~ while self._ws == None:
+                    #~ try:
+                        #~ url = http_client.fetch("http://localhost:%s/json" % debugport).body
+                        #~ self._ws = json.loads(url)[0]["webSocketDebuggerUrl"]
+                    #~ except Exception as e:
+                        #~ self._ws = None
 
     def wait(self):
-        self.__instance.wait()
+        self._p.wait()
+
+    def __del__(self): # really important !
+        self._p.kill()
+        if self.cacheFolderToRemove: shutil.rmtree(self.cacheFolderToRemove, ignore_errors=True)
+            
+    #~ def _com(self, payload: dict):
+        #~ """ https://chromedevtools.github.io/devtools-protocol/tot/Browser/#method-close """
+        #~ payload["id"] = 1
+        #~ r=json.loads(wsquery(self._ws,json.dumps(payload)))["result"]
+        #~ return r or True
+
+    #~ def focus(self): # not used
+        #~ return self._com(dict(method="Page.bringToFront"))
+
+    #~ def navigate(self, url): # not used
+        #~ return self._com(dict(method="Page.navigate", params={"url": url}))
 
     def exit(self):
-        self.__instance.kill()
+        #~ self._com(dict(method="Browser.close"))
+        self._p.kill()
 
 
-class ChromeAppCef:
-    def __init__(self, url, size=None, chromeArgs=None):  # chromeArgs is not used
+
+class CefApp:
+    def __init__(self, url, size=None, chromeArgs=None,lockPort=None):  # chromeArgs is not used
         import pkgutil
 
         assert pkgutil.find_loader("cefpython3"), "cefpython3 not available"
@@ -521,6 +607,10 @@ class ChromeAppCef:
                     devtools=True,
                 ),
             }
+            if lockPort:
+                settings["remote_debugging_port"]=lockPort
+                settings["cache_path"]= CHROMECACHE
+
             cef.Initialize(settings, {})
             b = cef.CreateBrowserSync(windowInfo, url=url)
 
@@ -583,6 +673,50 @@ async def doInit( instance ):
             self_init(  )
 
 
+def chromeBringToFront(port):
+    if not isFree("localhost", port):
+        http_client = tornado.httpclient.HTTPClient()
+        url = http_client.fetch("http://localhost:%s/json" % port).body
+        wsurl= json.loads(url)[0]["webSocketDebuggerUrl"]
+        wsquery(wsurl,json.dumps(dict(id=1,method="Page.bringToFront")))
+        return True        
+
+class LockPortFile:
+    def __init__(self,name):
+        self._file = os.path.join(CHROMECACHE,name,"lockport")
+
+    def bringToFront(self):
+        if os.path.isfile(self._file): # the file is here, perhaps it's running
+            with open(self._file,"r") as fid:
+                port=fid.read()
+            
+            if not isFree("localhost", int(port)): # if port is taken, perhaps it's running
+                http_client = tornado.httpclient.HTTPClient()
+                url = http_client.fetch("http://localhost:%s/json" % port).body
+                wsurl= json.loads(url)[0]["webSocketDebuggerUrl"]
+                print("*** ALREADY RUNNING")
+                wsquery(wsurl,json.dumps(dict(id=1,method="Page.bringToFront")))
+                return True
+
+    
+    def create(self) -> int:
+        if os.path.isfile(self._file):
+            os.unlink(self._file)
+        # find a freeport
+        port=9990
+        while not isFree("localhost", port):
+            port += 1
+
+        if not os.path.isdir( os.path.dirname(self._file)):
+            os.makedirs( os.path.dirname(self._file) )
+        with open(self._file,"w") as fid:
+            fid.write(str(port))
+
+        return port
+
+
+
+
 class Guy:
     _wsock=None     # when cloned and connected to a client/wsock (only the cloned instance set this)
 
@@ -627,18 +761,28 @@ class Guy:
         asyncio.ensure_future( doInit(self) )
 
 
-    def run(self,log=False,autoreload=False):
+    def run(self,log=False,autoreload=False,one=False,args=[]):
         """ Run the guy's app in a windowed env (one client)"""
         self._log=log
-        if log: handler.setLevel(logging.DEBUG)
+        if log: 
+            handler.setLevel(logging.DEBUG)
+            logger.setLevel(logging.DEBUG)
 
         if ISANDROID: #TODO: add executable for kivy/iOs mac/apple
             runAndroid(self)
         else:
+            lockPort=None
+            if one:
+                lp=LockPortFile(self._name)
+                if lp.bringToFront():
+                    return
+                else:
+                    lockPort = lp.create()
+            
             ws=WebServer( self, autoreload=autoreload )
             ws.start()
 
-            app=ChromeApp(ws.startPage,self.size)
+            app=ChromeApp(ws.startPage,self._name,self.size,lockPort=lockPort,chromeargs=args)
 
             def exit():
                 ws.exit()
@@ -657,15 +801,25 @@ class Guy:
 
         return self
 
-    def runCef(self,log=False,autoreload=False):
+    def runCef(self,log=False,autoreload=False,one=False):
         """ Run the guy's app in a windowed cefpython3 (one client)"""
         self._log=log
-        if log: handler.setLevel(logging.DEBUG)
+        if log: 
+            handler.setLevel(logging.DEBUG)
+            logger.setLevel(logging.DEBUG)
+
+        lockPort=None
+        if one:
+            lp=LockPortFile(self._name)
+            if lp.bringToFront():
+                return
+            else:
+                lockPort = lp.create()
 
         ws=WebServer( self, autoreload=autoreload )
         ws.start()
 
-        app=ChromeAppCef(ws.startPage,self.size)
+        app=CefApp(ws.startPage,self.size,lockPort=lockPort)
 
         tornado.autoreload.add_reload_hook(app.exit)
 
@@ -682,7 +836,9 @@ class Guy:
     def serve(self,port=8000,log=False,open=True,autoreload=False):
         """ Run the guy's app for multiple clients (web/server mode) """
         self._log=log
-        if log: handler.setLevel(logging.DEBUG)
+        if log: 
+            handler.setLevel(logging.DEBUG)
+            logger.setLevel(logging.DEBUG)
 
         ws=WebServer( self ,"0.0.0.0",port=port, autoreload=autoreload )
         ws.start()
@@ -765,7 +921,6 @@ class Guy:
         logger.debug("ROUTES: %s",routes)
         js = """
 document.addEventListener("DOMContentLoaded", function(event) {
-    %s
     %s
 },true)
 
@@ -992,7 +1147,6 @@ var self= {
 
 
 """ % (
-        size and "window.resizeTo(%s,%s);" % (size[0], size[1]) or "",
         'if(!document.title) document.title="%s";' % self._name,
         self._id, # for the socket
         "true" if self._log else "false",
@@ -1041,6 +1195,7 @@ var self= {
         ret= function(*args)
 
         if isinstance(ret,Guy):
+            print("*** DEPRECATED ***, will be removed in near future")
             ################################################################
             o=ret